pyproject.toml
[tool.poetry]
name = "aetherguild.net"
version = "1.0.0"
description = "Aether website"
authors = ["Tuomas Virtanen <katajakasa@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1"
django-imagekit = "^4.1.0"
django-redis = "^5.2.0"
django-crispy-forms = "^1.14.0"
django-silk = "^5.0.3"
django-recaptcha = "^3.0.0"
django-timezone-field = "^5.0"
django-csp = "^3.7"
django-csp-reports = "^1.8.1"
djangorestframework = "^3.14.0"
Pillow = "^9.4.0"
sentry-sdk = "^1.14.0"
argon2-cffi = "^21.2.0"
django-precise-bbcode = "^1.2.15"
celery = "^5.2.7"
requests = "^2.28.2"
uvicorn = "^0.20.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 109
target-version = ['py311']
